refactor: 💡 change the config file to @/common/constant.ts
all the config should manage by the center dir

BREAKING CHANGE: 🧨 none

✅ Closes: none
